refactor: split huge template.rs into folder with submodules
