+++ Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 authors = ["Ayush <ayush.agr254@gmail.com>", "Arjun <arjunmundamani@gmail.com>"]
 description = "A CLI tool for setting up pinocchio program project"
 repository = "https://github.com/a91y/pinoc"
-license = "MIT"
+license = "Apache-2.0"
 readme = "README.md"
 keywords = ["cli", "scaffold", "template", "solana", "pinocchio"]
 categories = ["command-line-utilities", "development-tools"]
